update bot to ts strict mode, replace puppeteer with node-canvas & minor improments

+++ packages/bot/src/SlashCommands/Music/LeaveCommand.ts
@@ -8,6 +8,13 @@ import { MusicPlayer } from '../../MusicPlayer/MusicPlayer';
 @injectable()
 export default class LeaveCommand extends AbstractMusicCommand {
   async doExecute(interaction: CommandInteraction, player: MusicPlayer): Promise<void> {
+    if (!interaction.inCachedGuild()) {
+      const answer = embedFactory(interaction.client, 'Error');
+      answer.setDescription('Command can not be executed inside DMs!');
+      await interaction.reply({ embeds: [answer], allowedMentions: { repliedUser: true }, ephemeral: true });
+      return;
+    }
+
     await MusicPlayerRepository.destroy(interaction.guild.id);
 
     const answer = embedFactory(interaction.client, 'Left the channel');
